Move RustString out of utils header
`utils` is more of an "unsorted pile of functions". RustString is pretty
fine as a stand-alone module.
