+++ include/ruststring.h
@@ -1,0 +1,26 @@
+#ifndef NEWSBOAT_RUSTSTRING_H_
+#define NEWSBOAT_RUSTSTRING_H_
+
+#include <string>
+
+namespace newsboat {
+
+class RustString {
+private:
+	char* str;
+
+public:
+	RustString() = delete;
+	RustString(const RustString&) = delete;
+	RustString(RustString&& rs) = delete;
+	RustString& operator=(RustString&& rs) noexcept = delete;
+
+	explicit RustString(char* ptr);
+	~RustString();
+
+	operator std::string();
+};
+
+}
+
+#endif /* NEWSBOAT_RUSTSTRING_H_ */
