+++ test/utils.cpp
@@ -1392,12 +1392,3 @@ TEST_CASE("mkdir_parents() doesn't care if the path ends in a slash or not",
 		check(path + "/");
 	}
 }
-
-TEST_CASE("RustString constructor and conversion to std::string", "[utils]")
-{
-	RustString foo(rs_get_string("foo"));
-	REQUIRE(std::string(foo) == std::string("foo"));
-
-	RustString bar(nullptr);
-	REQUIRE(std::string(bar) == std::string());
-}
